polymarket-arbitrage/requirements.txt
# Core dependencies
py-clob-client>=0.20.0
requests>=2.31.0
websocket-client>=1.6.0
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0

# Notifications
python-telegram-bot>=20.0

# Utilities
pyyaml>=6.0
colorlog>=6.7.0
schedule>=1.2.0

# API rate limiting
ratelimit>=2.2.1

# Optional: Web dashboard
flask>=3.0.0
plotly>=5.17.0
